Move some functions into mac_utils
